requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.10
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.7.0
async-timeout==5.0.1
attrs==21.2.0
Automat==20.2.0
Babel==2.8.0
bcrypt==3.2.0
bitsandbytes==0.44.1
blinker==1.9.0
certifi==2020.6.20
chardet==4.0.0
click==8.1.8
cloud-init==24.3.1
colorama==0.4.4
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
contourpy==1.3.1
cryptography==3.4.8
cycler==0.12.1
dbus-python==1.2.18
distro==1.7.0
distro-info==1.1+ubuntu0.2
exceptiongroup==1.2.2
faiss-cpu==1.9.0.post1
faiss-gpu==1.7.2
filelock==3.16.1
Flask==3.1.0
Flask-Cors==5.0.0
fonttools==4.55.0
frozenlist==1.5.0
fsspec==2024.10.0
gyp==0.1
h11==0.14.0
httpcore==1.0.7
httplib2==0.20.2
httpx==0.28.1
huggingface-hub==0.26.3
hyperlink==21.0.0
idna==3.3
imbalanced-learn==0.13.0
importlib-metadata==4.6.4
incremental==21.3.0
itsdangerous==2.2.0
jeepney==0.7.1
Jinja2==3.1.5
jiter==0.8.2
joblib==1.4.2
jsonpatch==1.32
jsonpointer==2.0
jsonschema==3.2.0
keyring==23.5.0
kiwisolver==1.4.7
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lightgbm==4.5.0
MarkupSafe==3.0.2
matplotlib==3.9.2
mlnx-tools===-5.8.0-
more-itertools==8.10.0
mpmath==1.3.0
multidict==6.1.0
netifaces==0.11.0
networkx==3.4.2
numpy==2.1.3
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
oauthlib==3.2.0
openai==1.58.1
packaging==21.3
pexpect==4.8.0
pillow==11.0.0
propcache==0.2.1
psutil==6.1.0
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.1
pydantic==2.10.4
pydantic_core==2.27.2
PyGObject==3.42.1
PyHamcrest==2.0.2
PyJWT==2.3.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyrsistent==0.18.1
pyserial==3.5
python-apt==2.4.0+ubuntu4
python-dateutil==2.9.0.post0
python-debian==0.1.43+ubuntu1.1
python-dotenv==1.0.1
python-magic==0.4.24
pytz==2022.1
PyYAML==5.4.1
regex==2024.11.6
requests==2.25.1
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
SecretStorage==3.3.1
sentence-transformers==3.3.1
service-identity==18.1.0
six==1.16.0
sklearn-compat==0.1.3
sniffio==1.3.1
sos==4.7.2
ssh-import-id==5.11
sympy==1.13.1
systemd-python==234
threadpoolctl==3.5.0
tokenizers==0.20.3
torch==2.5.1
torch-geometric==2.6.1
tqdm==4.67.1
transformers==4.46.3
triton==3.1.0
Twisted==22.1.0
typing_extensions==4.12.2
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.5
wadllib==1.3.6
Werkzeug==3.1.3
xkit==0.0.0
yarl==1.18.3
zipp==1.0.0
zope.interface==5.4.0
